fix(#7588): merge GlobalsSealTest into GlobalsTest to satisfy jtcop
